app/requirements.txt
pip-chill==1.0.3
psycopg2-binary==2.9.5
python-decouple==3.8
